// source/frontend/src/models/api/generated/ApiGen_Concepts_LeaseStakeholderType.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Concepts_CodeType } from './ApiGen_Concepts_CodeType';

// LINK: @backend/apimodels/Models/Concepts/Lease/LeaseStakeholderTypeModel.cs
export interface ApiGen_Concepts_LeaseStakeholderType extends ApiGen_Concepts_CodeType {
  isPayableRelated: boolean;
  isDisabled: boolean;
}